requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ChatterBot==1.2.8
chatterbot-corpus==1.2.2
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.3.0
Flask==3.1.2
Flask-Caching==2.3.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
frozenlist==1.8.0
h3==4.3.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mathparse==0.2.6
multidict==6.7.0
murmurhash==1.0.13
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
preshed==3.0.10
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
smart_open==7.5.0
spacy==3.8.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.1
thinc==8.3.8
timezonefinder==8.1.0
tomli==2.3.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.2
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0
